Move PointFileReaderWriter back to the core packages (used by MapData.java for reads)
